+++ README.md
@@ -473,7 +473,7 @@ This project is [MIT](https://github.com/woowacourse/java-subway-map-precourse/b
 
 ## ☑ 구현할 기능 목록
 
-- [ ] 메인
+- [x] 메인
   - [x] 선택지에 없는 입력일 시 에러메세지를 출력한다
   - [x] 사용자 입력에 따라 기능을 실행한다
-  - [ ] 사용자 입력이 Q인 경우 프로그램을 완전 종료한다
+  - [x] 사용자 입력이 Q인 경우 프로그램을 완전 종료한다
